backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.2
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
Pillow==10.1.0
django-cors-headers==4.3.1
boto3==1.35.0
django-storages==1.14.4